fixing plugin export
